Secure connection work for Prosit.
