+++ pwiz_tools/Skyline/Model/Prosit/Config/PrositConfig.cs
@@ -1,8 +1,29 @@
-﻿using System.Xml.Serialization;
+﻿/*
+ * Original author: Nicholas Shulman <nicksh .at. u.washington.edu>,
+ *                  MacCoss Lab, Department of Genome Sciences, UW
+ *
+ * Copyright 2019 University of Washington - Seattle, WA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+using System.Xml.Serialization;
 using Grpc.Core;
 
 namespace pwiz.Skyline.Model.Prosit.Config
 {
+    /// <summary>
+    /// Configuration of a Prosit server.
+    /// </summary>
     public class PrositConfig
     {
         public string Server { get; set; }
@@ -30,6 +51,9 @@ namespace pwiz.Skyline.Model.Prosit.Config
         }
 
 
+        /// <summary>
+        /// Read the configuration from the PrositConfig.xml embedded resource
+        /// </summary>
         public static PrositConfig GetPrositConfig()
         {
             using (var stream =
